refactor: SeoEditコンポーネントをadmin templateに統合
- SeoEditとSeoEditDialogコンポーネントを削除
- SEO設定機能をadmin/seo/template/index.tsxに直接統合
- 不要なコンポーネント階層を削除し、コードを簡素化
