--- frontend/src/app/admin/seo/template/SeoEdit/SeoEditDialog/styles.module.scss
@@ -1,39 +1,0 @@
-.form {
-    display: flex;
-    flex-direction: column;
-    gap: 1.5rem;
-    min-width: 400px;
-
-    @include media('sm') {
-        min-width: auto;
-    }
-}
-
-.field {
-    display: flex;
-    flex-direction: column;
-    gap: 0.5rem;
-}
-
-.label {
-    color: $text-color;
-    font-size: 1rem;
-    font-weight: bold;
-    margin-bottom: 0.5rem;
-}
-
-.logo-preview {
-    align-items: center;
-    display: flex;
-    justify-content: center;
-    margin-bottom: 1rem;
-}
-
-.logo-image {
-    border: 2px solid $light-dark-color;
-    border-radius: 50%;
-    height: 240px;
-    overflow: hidden;
-    position: relative;
-    width: 240px;
-}
