--- frontend/src/app/admin/seo/template/SeoEdit/styles.module.scss
@@ -1,57 +1,0 @@
-.seo-edit {
-    display: flex;
-    flex-direction: column;
-    gap: 3rem;
-}
-
-.section {
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
-}
-
-.section-title {
-    color: $title-primary-color;
-    font-size: 1.5rem;
-    font-weight: bold;
-
-    @include media('sm') {
-        font-size: 1.25rem;
-    }
-}
-
-.logo-section {
-    align-items: center;
-    display: flex;
-    gap: 2rem;
-    justify-content: center;
-
-    @include media('sm') {
-        flex-direction: column;
-        gap: 1rem;
-    }
-}
-
-.logo-preview {
-    flex-shrink: 0;
-}
-
-.logo-image {
-    border: 2px solid $light-dark-color;
-    border-radius: 50%;
-    height: 240px;
-    overflow: hidden;
-    position: relative;
-    width: 240px;
-}
-
-.description-section {
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
-}
-
-.edit-button {
-    align-self: center;
-    max-width: 120px;
-}
