+++ frontend/src/components/bases/Input/styles.module.scss
@@ -44,8 +44,9 @@
 
 .chip-container {
     position: absolute;
-    top: 44px;
+    top: 50%;
     right: 16px;
+    transform: translateY(-50%);
     z-index: 1;
 }
