+++ frontend/src/components/bases/SelectForm/styles.module.scss
@@ -18,7 +18,7 @@
 
 .chip-container {
     position: absolute;
-    top: 44px;
+    top: 50%;
     right: 16px;
     z-index: 1;
 }
